ai-orchestrator/requirements.txt
llama-cpp-python
pydantic
faiss-cpu
sentence-transformers
pyyaml
numpy
